Remove raw descriptors and two unused servlets.
We stopped processing server descriptors and linking raw descriptors
in the latest version of ExoneraTor. The idea was to simplify the
service and make it useful for the average user who doesn't care about
this level of detail. And it's a good opportunity to prevent the
database from growing even larger.

Implements #16595.