Fix query that took almost forever.
Queries that match one or more exit list entries took almost forever,
because avoided all the fine indexes we created for it.  The workaround is
to search by /24 of the requested IP address (or /48 in case of IPv6)
which we already have in the database.  We should want to do this anyway
to avoid making subsequent queries if there's no hit for the exact IP
address that was requested.

Fixes #17488.  Thanks, starlight, for all the feedback!