Use metrics-base as build environment.
The task 'war' creates a war-file for deployment in tomcat (as before),
the 'run' task is gone, and 'jar' creates an executable jar with main
class ExoneraTorDatabaseImporter as well as the source and javadoc jars.

Removed obsolete shell script.

Coverage is really low, but increasing it wasn't part of this task.

Implements task-21145.