Print out a warning if we're missing data.
We're now counting the hours for which we have statuses and exit
lists. More precisely, we're truncating consensus valid-after times
and exit scan times to the hour and counting how many distinct hours
we saw during the requested time. If we're missing 18 or more hours of
statuses or exit lists we're printing out a note in the summary.

Fixes #31071.