Improve logging of database importer.
This patch focuses on providing more context in log messages and unify
the use of log levels. It also adds counters and timestamps as far as
they are being used in log messages. And in very few cases it contains
minor changes like using try-with-resource when we're touching that
code anyway. But most other refactorings are left as future work.

Fixes the first part of #24542. The second part will be to improve
logging of the web server parts.